printers > printing for sympy convention.
