+++ sympy/physics/vector/__init__.py
@@ -1,10 +1,10 @@
 from .frame import ReferenceFrame, CoordinateSym
 from .dyadic import Dyadic
 from .vector import Vector
-from .printers import VectorStrPrinter, VectorLatexPrinter, \
-     VectorPrettyPrinter
+from .printing import (VectorStrPrinter, VectorLatexPrinter,
+                       VectorPrettyPrinter)
 from .point import Point
-from .functions import cross, dot, express, time_derivative, outer, \
-     time_derivative_printing, vprint, vsprint, vpprint, vlatex, \
-     kinematic_equations, get_motion_params, partial_velocity, \
-     dynamicsymbols
+from .functions import (cross, dot, express, time_derivative, outer,
+                        time_derivative_printing, vprint, vsprint, vpprint,
+                        vlatex, kinematic_equations, get_motion_params,
+                        partial_velocity, dynamicsymbols)
